Fix x64 version doesn't compile.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@3801 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qs/src/ui/textwindow.cpp
@@ -3875,7 +3875,7 @@ LRESULT qs::TextWindow::onPaint()
 	ObjectSelector<HPEN> penSelector(dc, hpenLink.get());
 	GdiObject<HPEN> hpenQuote[] = { 0, 0, 0 };
 	if (isLineQuote()) {
-		for (size_t n = 0; n < countof(hpenQuote); ++n)
+		for (unsigned int n = 0; n < countof(hpenQuote); ++n)
 			hpenQuote[n].reset(::CreatePen(PS_SOLID, 2,
 				TextWindowImpl::getAdjustedQuoteColor(pImpl_->crQuote_[0], n)));
 	}
